interrupts.rs: move ISRs out to interrupts.s
Leave in interrupts.rs only a dummy handler, which prints out the
interrupt number and an error code.

By default, all gates are dummy, which means that they are pointing to
fictional ISRs 256 and 257.  They call dummy_interrupt_handler(), which
prints out an interrupt number 256 or 257 (depending on the presence of
an error code).

The IDT static's constructor sets for each of the 32 exceptions a dummy
ISR that pushes onto the stack their own interrupt number and an error
code (0 if it's absent) and then calls dummy_interrupt_handler().