Add allocator.rs: implement GlobalAlloc
Each chunk (i.e. region of the heap) is preceded by a Tag struct aligned
at 2 bytes, which consists of two usize fields:

   1. address of the next tag | used bit (1 - used, 0 - free chunk);
   2. alignment requested by the caller.

(Here | is for logical OR.)

There is always a tag at the heap beginning.  So by knowing only the
heap region start, one can obtain all the tags.

To mark the end of the heap, at the address heap_end - sizeof::<Tag>()
there's a tag with the following fields:

   1. 0, meaning the next tag address is null and the chunk is not used;
   2. alignment is 1 byte (just because it has to be a power of two).

Now, because GlobalAlloc::dealloc() accepts the chunk start address and
not the tag address, it has to determine where the latter is to alter
it (to free the chunk).  Given that the chunk start is aligned, it can
be anywhere.

To solve this problem, the area between the tag and the chunk start is
filled with the byte 0xFF before returning from GlobalAlloc::alloc().
Because the last field of Tag, Tag::align: usize, is a power of two, it
has only one of the 32 (or 64) bits set, and its 4 (or 8) bytes can
never be 0xFF.