Add Process struct and Round Robin scheduling
After the kernel finishes initializing its vital stuff, that is,
resource managers, the PIT, etc., it creates a usermode init process and
jumps into usermode using jump_into_usermode (scheduler.s) with the
address of the init function passed as an argument.

The init function (which is totally bare) then is interrupted by an IRQ0
handler which runs the scheduler.  At one moment the latter decides to
switch tasks using the Round Robin scheme (i.e. simply go to the next
process in the list).  It does so by:

  1. saving the context of the current task on its kernel heap (that's
     where %esp points at because the scheduler code is run in the IRQ0
     handler),

  2. making %esp point at the next task's kernel heap (where it always
     expects the previously saved context to be),

  3. restoring the next task's context by popping everything (except
     eip) from its kernel stack,

  4. doing a `ret` to where the previous task was interrupted (restoring
     eip), thus fully restoring its context.

The interesting thing is creating a new task.  Besides allocating a
kernel stack for it, one has to push a new task's context as if it was
interrupted by something.  Then, the function that saved %eip points at
must enable interrupts as the task switching was done in the IRQ0
handler with interrupts disabled.  The task then can be actually
interrupted by another context switch, thus implementing multitasking.

Also, the IRQ0 handler creates a few processes to test the scheduler.