Add PCI support
Each PCI host controller is represented by a corresponding function of
device 0 on bus 0.  If a function exists, Pci::enumerate() scans the
host bus by calling Bus::new(function_num) and then pushes the resulting
Bus struct to Pci::host_buses.

Each bus has 32 devices, the types of which are determined based on the
header type register of function 0, - although this may not be always
right, I think.  (Configuration space access is done by port I/O, memory
map access which was introduced in PCI Express is not implemented.)  The
kernel understands two kinds of devices: an actual Device and a
PCI-to-PCI-bridge.  If it's the former, a Device struct is created and
pushed to Bus::devices.  If it's the latter, the secondary bus num. is
extracted from function 0 configurtaion space and Bus::new() is called
for that bus, the result of which call, a Bus struct, is pushed to
Bus::secondary_buses.

If a device says it has only one function, that function (it must have
number 0) is pushed to Device::functions.  If it is multifunctional,
functions 0 to 7 are pushed, regardless whether the function exists
(i.e vendor ID is not 0xFFFF) or not.