Add ATA driver
The empty struct Driver is wrapped into a static Mutex.  This ensures
that its methods will never be executed simultaneously.

Reading is done using port I/O and LBA addressing.  After sending a read
command, the driver checks for errors and waits until the controller is
ready to send data.  Then it allocates a buffer, copies the data from
the port to that buffer and returns it as a slice.

Only channel 0 slave hard drive is supported.  No writing routine yet.