ext2.rs: don't rely on packed(N) where N > 4
For some reason packed(32) becomes packed(4), so it does not provide
needed alignment.

Also, iterating over the block group descriptor table is not possible
using the current way.  It relies on the fact that the BGD table, which
is provided as an argument, is aligned at 32 bytes, which is obviously
wrong.  So just hard-code the BGD table entry size into the for loop.