Scheduler: add stop_scheduling() and keep_scheduling()
These methods allow tasks do their critical stuff without being
interrupted (e.g. by an IRQ0 -> a task switch).  They are built upon an
atomic counter which is equal to the number of such tasks, and when it's
set to zero, interrupts are enabled back.

It's important to note that interrupts are enabled even if initially
(i.e. before the first call to stop_scheduling) they were disabled.  So
these methods should be used only by tasks and not the kernel.