ata.rs: expose Drives, not Bus, to DISKS
Previously disk::DISKS had a rw_interface associated with ata::Bus which
forced the kernel to use either the master or the slave drive.  Now it's
associated with ata::Drive and it became possible to use both drives.

To achieve this a bus field was added to Drive which is conceptually a
shared pointer to Bus.  Bus itself does not know about its Drives and it
doesn't need to.  Each of the Drives is aware which one of the two it
is, thus configures the bus appropriately and does its reading/writing
stuff.