process: support opening files
Each new process is given an empty vector of file handles.  A file
handle is an fs::Node with an offset counter and a buffer for the file
contents.  Opening a file is done via Process::open_file_by_node() which
returns some file descriptor, an index in the aforementioned vector that
corresponds to the opened file.