Add devfs
This required to add support of mounting.  Everything that implements
fs::Mountable can be mounted using fs::Node::mount_on_child() on an
existing empty directory.

In /dev there are only block devices for the disks known to the kernel.
As a test the processes open /dev/disk0 and read at an arbitrary byte
offset.

Use more doc comments.